chore: restructure database
