+++ crates/biome_html_analyze/src/lib.rs
@@ -16,7 +16,7 @@ use crate::suppression_action::HtmlSuppressionAction;
 /// Services available to HTML lint rules.
 #[derive(Default)]
 pub struct HtmlAnalyzerServices {
-    pub module_db: Option<Arc<dyn ModuleDb>>,
+    pub module_db: Option<ProjectDatabase>,
     pub project_layout: Option<Arc<ProjectLayout>>,
 }
 
@@ -37,7 +37,7 @@ use biome_aria::AriaRoles;
 use biome_deserialize::TextRange;
 use biome_diagnostics::Error;
 use biome_html_syntax::{HtmlFileSource, HtmlLanguage};
-use biome_module_graph::ModuleDb;
+use biome_module_graph::ProjectDatabase;
 use biome_project_layout::ProjectLayout;
 use biome_suppression::{SuppressionDiagnostic, parse_suppression_comment};
 use std::ops::Deref;
