+++ crates/biome_module_graph/src/db/mod.rs
@@ -1,3 +1,4 @@
 mod impls;
 pub mod inputs;
+pub mod project_database;
 pub mod queries;
