+++ crates/biome_service/src/workspace/db.rs
@@ -1,10 +1,6 @@
 use crate::WorkspaceError;
-use biome_db::Db;
-use biome_module_graph::ModuleInfoKind;
-use biome_module_graph::{ModuleDb, ModuleInfo, PathInfoCache};
-use camino::{Utf8Path, Utf8PathBuf};
-use papaya::HashMap;
-use salsa::Storage;
+pub use biome_module_graph::ProjectDatabase;
+use biome_module_graph::PathInfoCache;
 use std::sync::{Mutex, MutexGuard};
 
 /// Represents the state of the database in the workspace.
@@ -19,39 +15,5 @@ impl DbState {
         self.db
             .lock()
             .map_err(|_| WorkspaceError::db_lock_poisoned())
-    }
-}
-
-#[salsa::db]
-#[derive(Default, Clone)]
-pub struct ProjectDatabase {
-    pub(crate) modules: HashMap<Utf8PathBuf, ModuleInfo>,
-    storage: Storage<ProjectDatabase>,
-}
-
-impl ProjectDatabase {
-    pub fn insert_module(&self, path: Utf8PathBuf, module: ModuleInfo) {
-        self.modules.pin().insert(path, module);
-    }
-}
-
-#[salsa::db]
-impl salsa::Database for ProjectDatabase {}
-
-#[salsa::db]
-impl Db for ProjectDatabase {}
-
-#[salsa::db]
-impl ModuleDb for ProjectDatabase {
-    fn module_for_path(&self, path: &Utf8Path) -> Option<ModuleInfo> {
-        self.modules.pin().get(path).copied()
-    }
-
-    fn for_each_module(&self, f: &mut dyn FnMut(&Utf8Path, &ModuleInfoKind)) {
-        let modules = self.modules.pin();
-        for (path, &module_info) in modules.iter() {
-            let kind = module_info.kind(self);
-            f(path.as_path(), &kind);
-        }
     }
 }
